Added a testnet and mainnet deployment stage
Added in a deployment section that distinguished between Mumbai testnet and Polygon mainnet. I work at Alchemy and based on our own developers' questions and ask for tutorials, thought the rewritten section would be much more clear to follow for developers in terms of connecting your Metamask wallet and deploying on testnet and mainnet. Lmk of any questions, thanks!