Move Alchemy's Node page to the "Setup Archive Node" section
Hey there -- moving this page to be a visible subsection in the overall archive page instead of by itself -- it should be an option on this page (https://wiki.polygon.technology/docs/category/setup-archive-node) and be a fourth subsection called "Using Alchemy". 

Context -- this "Setup Archive Node" section should be where all the externally hosted node providers should live -- along with Binaries / Ansible, Erigon Client, and NodeReal, the Alchemy node quickstart page should be found in this section for developers who are looking for this.  Pls let me know of any questions, thanks!